refactor: 소셜 토큰 관련 클래스를 users.auth 패키지로 이동
