+++ src/main/java/com/bookspot/users/application/UserTokenService.java
@@ -4,12 +4,10 @@ import com.bookspot.global.auth.JwtProvider;
 import com.bookspot.global.auth.dto.GeneratedToken;
 import com.bookspot.users.application.helper.SocialTokenVerifierSelector;
 import com.bookspot.users.domain.OAuthProvider;
-import com.bookspot.users.domain.SocialTokenDetail;
-import com.bookspot.users.domain.SocialTokenVerifier;
-import com.bookspot.users.infra.token.google.GoogleTokenVerifier;
+import com.bookspot.users.domain.auth.SocialTokenDetail;
+import com.bookspot.users.domain.auth.SocialTokenVerifier;
 import com.bookspot.users.infra.token.naver.NaverTokenVerifier;
 import com.bookspot.users.presentation.UserTokenResponse;
-import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
