+++ src/main/java/com/bookspot/users/domain/auth/SocialTokenDetail.java
@@ -1,4 +1,4 @@
-package com.bookspot.users.domain;
+package com.bookspot.users.domain.auth;
 
 public record SocialTokenDetail(String subjectId, String email) {
 }
