+++ src/main/java/com/bookspot/users/domain/auth/SocialTokenVerifier.java
@@ -1,4 +1,6 @@
-package com.bookspot.users.domain;
+package com.bookspot.users.domain.auth;
+
+import com.bookspot.users.domain.OAuthProvider;
 
 public interface SocialTokenVerifier {
     SocialTokenDetail verifyToken(String idTokenString);
